fix dependency
