+++ ci/scripts/intel_copyright_files.txt
@@ -27,6 +27,7 @@ cpp/src/cider/exec/nextgen/*
 cpp/src/cider/exec/processor/*
 cpp/src/cider/exec/operator/join/*
 cpp/src/cider/type/plan/*Expr.cpp
+cpp/src/cider/type/plan/ExprType.h
 cpp/src/cider/type/plan/InValues.cpp
 cpp/src/cider/exec/plan/lookup/FunctionLookupEngine.*
 cpp/src/cider/exec/plan/lookup/SubstraitFunctionCiderMappings.h
