+++ cpp/src/cider-velox/test/CMakeLists.txt
@@ -103,6 +103,7 @@ set(VELOX_PLUGIN_TEST_LIBS
     ${LLVM_LIB}
     cider
     cider_processor
+    cider_plan_substrait
     cider_plan_parser
     velox_plan_transformer
     cider_plan_transformer
