+++ cpp/src/cider/exec/plan/lookup/FunctionLookupEngine.cpp
@@ -21,11 +21,12 @@
 
 #include "exec/plan/lookup/FunctionLookupEngine.h"
 #include "cider/CiderException.h"
+#include "util/StringTransform.h"
 
 FunctionLookupEnginePtrMap FunctionLookupEngine::function_lookup_engine_ptr_map_ = {};
 
 std::mutex FunctionLookupEngine::s_mutex_;
-std::string FunctionLookupEngine::data_path_ = ""; // NOLINT
+std::string FunctionLookupEngine::data_path_ = "";  // NOLINT
 
 FunctionLookupEnginePtr FunctionLookupEngine::getInstance(
     const PlatformType from_platform) {
