+++ cpp/src/cider/exec/processor/CMakeLists.txt
@@ -22,5 +22,6 @@ set(PROCESSOR_SOURCE
     JoinHandler.cpp DefaultJoinHashTableBuilder.cpp ExpressionEvaluator.cpp)
 
 add_library(cider_processor STATIC ${PROCESSOR_SOURCE})
-target_link_libraries(cider_processor cider_plan_substrait cider_plan_parser
-                      nextgen cider_hashtable_join)
+target_link_libraries(cider_processor cider)
+# target_link_libraries(cider_processor cider cider_plan_substrait
+# cider_plan_parser nextgen cider_hashtable_join)
